+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,8 @@
 #### 🐞 Fixes
 
 - Fixed an issue where glob task outputs were not invalidating a previous build.
+- Fixed an issue where changing inputs would not mark a task as affected, when a moon workspace is
+  nested within a repository.
 
 ## 1.7.2